Break down MainMediaImage into components
